+++ tests/test_ci_contract.py
@@ -19,8 +19,6 @@ class CiContractTests(unittest.TestCase):
             "bash -n install.sh",
             "./bin/writing-master learn --help",
             "./bin/writing-master research --help",
-            "./bin/writing-master voice --help",
-            "./bin/writing-master voice list --json",
             "./bin/writing-master handoff --help",
             "python -m build",
         ):